requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==4.2.8
django-bootstrap-icons==0.8.7
django-bootstrap-v5==1.0.11
django-ckeditor==6.7.0
django-js-asset==2.2.0
idna==3.6
Pillow==10.1.0
psycopg==3.1.16
psycopg-binary==3.1.16
psycopg-pool==3.2.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
